Struct jit::CompiledFunction

pub struct CompiledFunction<'a> {
    // some fields omitted
}

A function which has already been compiled from an UncompiledFunction. This can be called but not added to.

A function persists for the lifetime of its containing context. This is a function which has already been compiled and is now in executable form.

Methods

impl<'a> CompiledFunction<'a>

fn with<A, R, F: FnOnce( extern "C" fn(A) -> R)>(self, cb: F)

Run a closure with the compiled function as an argument