Struct jit::Context

pub struct Context<T = ()> {
    // some fields omitted
}

Holds all of the functions you have built and compiled. There can be multiple, but normally there is only one.

Methods

impl<T = ()> Context<T>

fn new() -> Context<T>

Create a new JIT Context

fn build<'a, R, F: FnOnce(Builder<T>) -> R>(&'a mut self, cb: F) -> R

Lock the context so you can safely generate IR

fn build_func<'a, F: FnOnce(&UncompiledFunction<'a>)>(&'a mut self, signature: &Ty, cb: F) -> CompiledFunction<'a>

Lock the context so you can safely generate IR in a new function on the context which is compiled for you

fn functions(&self) -> Functions

Iterate through the functions contained inside this context