Struct jit::Ty

pub struct Ty(_);

A reference to an object that represents a native system type . This represents a basic system type, be it a primitive, a struct, a union, a pointer, or a function signature. The library uses this information to lay out values in memory.

Methods

impl Ty

fn get_size(&self) -> usize

Get the size of this type in bytes.

fn get_alignment(&self) -> usize

Get the alignment of this type in bytes.

fn get_kind(&self) -> TypeKind

Get a value that indicates the kind of this type. This allows callers to quickly classify a type to determine how it should be handled further.

fn get_ref(&self) -> Option<&Ty>

Get the type that is referred to by this pointer type.

fn get_return(&self) -> Option<&Ty>

Get the type returned by this function type.

fn set_names(&mut self, names: &[&str])

Set the field or parameter names of this type.

fn fields(&self) -> Fields

Iterator over the type's fields

fn params(&self) -> Params

Iterator over the function signature's parameters

fn get_field(&self, name: &str) -> Field

Find the field/parameter index for a particular name.

fn is_primitive(&self) -> bool

Check if this is a pointer

fn is_struct(&self) -> bool

Check if this is a struct

fn is_union(&self) -> bool

Check if this is a union

fn is_signature(&self) -> bool

Check if this is a signature

fn is_pointer(&self) -> bool

Check if this is a pointer

fn is_tagged(&self) -> bool

Check if this is tagged