Struct jit::UncompiledFunction

pub struct UncompiledFunction<'a> {
    // some fields omitted
}

A function which has not been compiled yet, so it can have instructions added to it.

A function persists for the lifetime of its containing context. This represents the function in the "building" state, where the user constructs instructions that represents the function body. Once the build process is complete, the user calls function.compile() to convert it into its executable form.

Methods

impl<'a> UncompiledFunction<'a>

fn new(context: &'a Builder, signature: Type) -> UncompiledFunction<'a>

Create a new function block and associate it with a JIT context. It is recommended that you call Function::new and function.compile() in the closure you give to context.build(...).

This will protect the JIT's internal data structures within a multi-threaded environment.

fn new_nested(context: &'a Builder, signature: Type, parent: &'a UncompiledFunction<'a>) -> UncompiledFunction<'a>

Create a new function block and associate it with a JIT context. In addition, this function is nested inside the specified parent function and is able to access its parent's (and grandparent's) local variables.

The front end is responsible for ensuring that the nested function can never be called by anyone except its parent and sibling functions. The front end is also responsible for ensuring that the nested function is compiled before its parent.

fn insn_convert(&self, v: &Value<'a>, t: Type, overflow_check: bool) -> Value<'a>

Make an instruction that converts the value to the type given

fn insn_of<T: Compile>(&self, val: &T) -> Value<'a>

Make an instructional representation of a Rust value

fn insn_uses_catcher(&self)

Notify the function building process that this function has a catch block in it. This must be called before any code that is part of a try block

fn insn_throw(&self, retval: &Value<'a>)

Throw an exception from the function with the value given

fn insn_return(&self, retval: &Value<'a>)

Return from the function with the value given

fn insn_default_return(&self)

Return from the function

fn insn_mul(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that multiplies the values

fn insn_mul_ovf(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that multiplies the values and throws upon overflow

fn insn_add(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that adds the values

fn insn_add_ovf(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that adds the values and throws upon overflow

fn insn_sub(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that subtracts the second value from the first

fn insn_sub_ovf(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that subtracts the second value from the first and throws upon overflow

fn insn_div(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that divides the first number by the second

fn insn_rem(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that finds the remainder when the first number is divided by the second

fn insn_leq(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that checks if the first value is lower than or equal to the second

fn insn_geq(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that checks if the first value is greater than or equal to the second

fn insn_lt(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that checks if the first value is lower than the second

fn insn_gt(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that checks if the first value is greater than the second

fn insn_eq(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that checks if the values are equal

fn insn_neq(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that checks if the values are not equal

fn insn_and(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that performs a bitwise and on the two values

fn insn_or(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that performs a bitwise or on the two values

fn insn_xor(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that performs a bitwise xor on the two values

fn insn_not(&self, value: &Value<'a>) -> Value<'a>

Make an instruction that performs a bitwise not on the two values

fn insn_shl(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that performs a left bitwise shift on the first value by the second value

fn insn_shr(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that performs a right bitwise shift on the first value by the second value

fn insn_ushr(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that performs a right bitwise shift on the first value by the second value

fn insn_neg(&self, value: &Value<'a>) -> Value<'a>

Make an instruction that performs a bitwise negate on the value

fn insn_dup(&self, value: &Value<'a>) -> Value<'a>

Make an instruction that duplicates the value given

fn insn_load(&self, src: &Value<'a>) -> Value<'a>

Make an instruction that loads a value from a src value

fn insn_load_relative(&self, src: &Value<'a>, offset: int, ty: Type) -> Value<'a>

Make an instruction that loads a value from a src value

fn insn_store(&self, dest: &Value<'a>, src: &Value<'a>)

Make an instruction that stores a value at a destination value

fn insn_store_relative(&self, dest: &Value<'a>, offset: int, src: &Value<'a>)

Make an instruction that stores a value a certain offset away from a destination value

fn insn_label(&self, label: &mut Label<'a>)

Make an instruction that sets a label

fn insn_branch(&self, label: &mut Label<'a>)

Make an instruction that branches to a certain label

fn insn_branch_if(&self, value: &Value<'a>, label: &mut Label<'a>)

Make an instruction that branches to a certain label if the value is true

fn insn_branch_if_not(&self, value: &Value<'a>, label: &mut Label<'a>)

Make an instruction that branches to a certain label if the value is false

fn insn_jump_table(&self, value: &Value<'a>, labels: &mut [Label<'a>])

Make an instruction that branches to a label in the table

fn insn_acos(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that gets the inverse cosine of the number given

fn insn_asin(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that gets the inverse sine of the number given

fn insn_atan(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that gets the inverse tangent of the number given

fn insn_atan2(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that gets the inverse tangent of the numbers given

fn insn_ceil(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that finds the nearest integer above a number

fn insn_cos(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that gets the consine of the number given

fn insn_cosh(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that gets the hyperbolic consine of the number given

fn insn_exp(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that gets the natural logarithm rased to the power of the number

fn insn_floor(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that finds the nearest integer below a number

fn insn_log(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that gets the natural logarithm of the number

fn insn_log10(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that gets the base 10 logarithm of the number

fn insn_pow(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction the gets the result of raising the first value to the power of the second value

fn insn_rint(&self, v: &Value<'a>) -> Value<'a>

Make an instruction the gets the result of rounding the value to the nearest integer

fn insn_round(&self, v: &Value<'a>) -> Value<'a>

Make an instruction the gets the result of rounding the value to the nearest integer

fn insn_sin(&self, v: &Value<'a>) -> Value<'a>

Make an instruction the gets the sine of the number

fn insn_sinh(&self, v: &Value<'a>) -> Value<'a>

Make an instruction the gets the hyperbolic sine of the number

fn insn_sqrt(&self, v: &Value<'a>) -> Value<'a>

Make an instruction the gets the square root of a number

fn insn_tan(&self, v: &Value<'a>) -> Value<'a>

Make an instruction the gets the tangent of a number

fn insn_tanh(&self, v: &Value<'a>) -> Value<'a>

Make an instruction the gets the hyperbolic tangent of a number

fn insn_trunc(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that truncates the value

fn insn_is_nan(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that checks if the number is NaN

fn insn_is_finite(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that checks if the number is finite

fn insn_is_inf(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that checks if the number is infinite

fn insn_abs(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that gets the absolute value of a number

fn insn_min(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that gets the smallest of two numbers

fn insn_max(&self, v1: &Value<'a>, v2: &Value<'a>) -> Value<'a>

Make an instruction that gets the biggest of two numbers

fn insn_sign(&self, v: &Value<'a>) -> Value<'a>

Make an instruction that gets the sign of a number

fn insn_call<S: ToCStr, F: Function<'a>>(&self, name: Option<S>, func: &F, sig: Option<Type>, args: &mut [&Value<'a>], flags: CallFlags) -> Value<'a>

Call the function, which may or may not be translated yet

fn insn_call_indirect(&self, func: &Value<'a>, signature: Type, args: &mut [&Value<'a>], flags: CallFlags) -> Value<'a>

Make an instruction that calls a function that has the signature given with some arguments through a pointer to the fucntion

fn insn_call_native0<R, S: ToCStr>(&self, name: Option<S>, native_func: extern "C" fn() -> R, signature: Type, flags: CallFlags) -> Value<'a>

Make an instruction that calls a Rust function that has the signature given with no arguments and expects a return value

fn insn_call_native1<A, R, S: ToCStr>(&self, name: Option<S>, native_func: extern "C" fn(A) -> R, signature: Type, args: [&Value<'a>, ..1], flags: CallFlags) -> Value<'a>

Make an instruction that calls a Rust function that has the signature given with a single argument and expects a return value

fn insn_call_native2<A, B, R, S: ToCStr>(&self, name: Option<S>, native_func: extern "C" fn(A, B) -> R, signature: Type, args: [&Value<'a>, ..2], flags: CallFlags) -> Value<'a>

Make an instruction that calls a Rust function that has the signature given with two arguments and expects a return value

fn insn_call_native3<A, B, C, R, S: ToCStr>(&self, name: Option<S>, native_func: extern "C" fn(A, B, C) -> R, signature: Type, args: [&Value<'a>, ..3], flags: CallFlags) -> Value<'a>

Make an instruction that calls a Rust function that has the signature given with three arguments and expects a return value

fn insn_call_native4<A, B, C, D, R, S: ToCStr>(&self, name: Option<S>, native_func: extern "C" fn(A, B, C, D) -> R, signature: Type, args: [&Value<'a>, ..4], flags: CallFlags) -> Value<'a>

Make an instruction that calls a Rust function that has the signature given with four arguments and expects a return value

fn insn_alloca(&self, size: &Value<'a>) -> Value<'a>

Make an instruction that allocates some space

fn insn_address_of(&self, value: &Value<'a>) -> Value<'a>

Make an instruction that gets the address of a value

fn insn_if<B: FnOnce()>(&self, cond: &Value<'a>, block: B)

Make instructions to run the block if the condition is met

fn insn_if_not<B: FnOnce()>(&self, cond: &Value<'a>, block: B)

Make instructions to run the block if the condition is not met

fn insn_loop<B: FnOnce()>(&self, block: B)

Make instructions to run the block forever

fn insn_loop_while<C: FnOnce() -> Value<'a>, B: FnOnce()>(&self, cond: C, block: B)

Make instructions to run the block and continue running it so long as the condition is met

fn insn_while<C: FnOnce() -> Value<'a>, B: FnOnce()>(&self, cond: C, block: B)

Make instructions to run the block and continue running it so long as the condition is met

fn set_optimization_level(&self, level: c_uint)

Set the optimization level of the function, where the bigger the level, the more effort should be spent optimising

fn get_max_optimization_level() -> c_uint

Get the max optimization level

fn set_recompilable(&self)

Make this function a candidate for recompilation

fn compile(self) -> CompiledFunction<'a>

Compile the function

fn compile_with<A, R, F: FnOnce( extern "C" fn(A) -> R)>(self, cb: F) -> CompiledFunction<'a>

Compile the function into a closure directly