Struct jit::WriteElf

pub struct WriteElf {
    // some fields omitted
}

An ELF binary reader

Methods

impl WriteElf

fn new<S: ToCStr>(lib_name: S) -> WriteElf

Create a new ELF binary reader

fn write<S: ToCStr>(&self, filename: S) -> bool

Write to the filename given (not implemented by LibJIT yet, so there's no point to this yet but I'm sure GNU will hear the people sing the songs of angry men soon enough)

fn add_function<S: ToCStr>(&self, func: &CompiledFunction, name: S) -> bool

Add a function to the ELF

fn add_needed<S: ToCStr>(&self, lib_name: S) -> bool

Add a dependency to the ELF

impl WriteElf

fn new<S: ToCStr>(lib_name: S) -> WriteElf

Create a new ELF binary reader

fn write<S: ToCStr>(&self, filename: S) -> bool

Write to the filename given (not implemented by LibJIT yet, so there's no point to this yet but I'm sure GNU will hear the people sing the songs of angry men soon enough)

fn add_function<S: ToCStr>(&self, func: &CompiledFunction, name: S) -> bool

Add a function to the ELF

fn add_needed<S: ToCStr>(&self, lib_name: S) -> bool

Add a dependency to the ELF