Struct jit::Type

pub struct Type {
    // some fields omitted
}

An object that represents a native system type. Each Type represents a basic system type, be it a primitive, a struct, a union, a pointer, or a function signature. The library uses this information to lay out values in memory. Types are not attached to a context so they are reference-counted by LibJIT, so internally they are represented as Rc<TypeData>.

Methods

impl Type

fn new_signature(abi: Abi, return_type: Type, params: &mut [Type]) -> Type

Create a type descriptor for a function signature.

fn new_struct(fields: &mut [Type]) -> Type

Create a type descriptor for a structure.

fn new_union(fields: &mut [Type]) -> Type

Create a type descriptor for a union.

fn new_pointer(pointee: Type) -> Type

Create a type descriptor for a pointer to another type.

fn get_size(&self) -> usize

Get the size of this type in bytes.

fn get_alignment(&self) -> usize

Get the alignment of this type in bytes.

fn get_kind(&self) -> TypeKind

Get a value that indicates the kind of this type. This allows callers to quickly classify a type to determine how it should be handled further.

fn get_ref(&self) -> Option<Type>

Get the type that is referred to by this pointer type.

fn get_return(&self) -> Option<Type>

Get the type returned by this function type.

fn with_names(self, names: &[&str]) -> Type

Set the field or parameter names of this type.

fn fields<'a>(&'a self) -> Fields<'a>

Iterator over the type's fields

fn params<'a>(&'a self) -> Params<'a>

Iterator over the function signature's parameters

fn get_field<'a>(&'a self, name: &str) -> Field<'a>

Find the field/parameter index for a particular name.

fn is_primitive(&self) -> bool

Check if this is a pointer

fn is_struct(&self) -> bool

Check if this is a struct

fn is_union(&self) -> bool

Check if this is a union

fn is_signature(&self) -> bool

Check if this is a signature

fn is_pointer(&self) -> bool

Check if this is a pointer

fn is_tagged(&self) -> bool

Check if this is tagged