Struct jit::Type

pub struct Type {
    // some fields omitted
}

An owned object that represents a native system type.

Each Type represents a basic system type, be it a primitive, a struct, a union, a pointer, or a function signature. The library uses this information to lay out values in memory.

Types are not attached to a context so they are reference-counted by LibJIT, so internally they are represented as Rc<Ty>.

Methods

impl Type

fn new_signature(abi: Abi, return_type: &Ty, params: &mut [&Ty]) -> Type

Create a type descriptor for a function signature.

fn new_struct(fields: &mut [&Ty]) -> Type

Create a type descriptor for a structure.

fn new_union(fields: &mut [&Ty]) -> Type

Create a type descriptor for a union.

fn new_pointer(pointee: &Ty) -> Type

Create a type descriptor for a pointer to another type.